gateway proxy split into files
